Move form_struct to its own file
And move specific interactions to html file for better locality.
Additionally, add macro_export instead of pub use.
